Extract SVG materialization helpers
